+++ R/provider.R
@@ -69,7 +69,7 @@ value_turn <- new_generic("value_turn", "provider")
 as_json <- new_generic("as_json", c("provider", "x"))
 
 method(as_json, list(Provider, class_list)) <- function(provider, x) {
-  lapply(x, as_json, provider = provider)
+  compact(lapply(x, as_json, provider = provider))
 }
 
 method(as_json, list(Provider, ContentJson)) <- function(provider, x) {