+++ tests/testthat/test-provider-openrouter.R
@@ -1,0 +1,53 @@
+# Getting started --------------------------------------------------------
+
+test_that("can make simple request", {
+  chat <- chat_openrouter_test("Be as terse as possible; no punctuation")
+  resp <- chat$chat("What is 1 + 1?", echo = FALSE)
+  expect_match(resp, "2")
+  expect_true(all(chat$last_turn()@tokens >= 1))
+})
+
+test_that("can make simple streaming request", {
+  chat <- chat_openrouter_test("Be as terse as possible; no punctuation")
+  resp <- coro::collect(chat$stream("What is 1 + 1?"))
+  expect_match(paste0(unlist(resp), collapse = ""), "2")
+})
+
+test_that("handles errors", {
+  chat <- chat_openrouter_test(api_args = list(temperature = "hot"))
+  expect_snapshot(error = TRUE, {
+    chat$chat("What is 1 + 1?", echo = FALSE)
+    chat$chat("What is 1 + 1?", echo = TRUE)
+  })
+})
+
+# Common provider interface -----------------------------------------------
+
+test_that("respects turns interface", {
+  chat_fun <- chat_openrouter_test
+
+  test_turns_system(chat_fun)
+  test_turns_existing(chat_fun)
+})
+
+test_that("all tool variations work", {
+  chat_fun <- chat_openrouter_test
+
+  test_tools_simple(chat_fun)
+  test_tools_async(chat_fun)
+  test_tools_parallel(chat_fun)
+  test_tools_sequential(chat_fun, total_calls = 6)
+})
+
+test_that("can extract data", {
+  chat_fun <- chat_openrouter_test
+
+  test_data_extraction(chat_fun)
+})
+
+test_that("can use images", {
+  chat_fun <- chat_openrouter_test
+
+  test_images_inline(chat_fun)
+  test_images_remote(chat_fun)
+})